+++ quarkus-petclinic-customers-service/src/main/java/io/quarkus/samples/petclinic/customers/web/PetDetails.java
@@ -26,7 +26,7 @@ import org.springframework.format.annotation.DateTimeFormat;
  */
 class PetDetails {
 
-    public Integer id;
+    public Long id;
 
     public String name;
